# api/openapi.yaml
openapi: 3.1.0
info:
  title: Hydra Private Chat API
  version: 0.0.1
  description: |
    v0 contract-first API spec for Hydra backend.
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      summary: Health check
      operationId: getHealth
      responses:
        '200':
          description: Service health
          content:
            application/json:
              schema:
                type: object
                required: [ok, time]
                properties:
                  ok:
                    type: boolean
                  time:
                    type: string
                    format: date-time

  /users/register:
    post:
      summary: Register a new user key bundle
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Signature verification failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Nonce reuse or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /users/{id}/keys:
    get:
      summary: Get a user's public key bundle
      operationId: getUserKeys
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: user_id
      responses:
        '200':
          description: Public keys found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserKeysResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /messages:
    post:
      summary: Send encrypted message envelope
      operationId: postMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageEnvelopeRequest'
      responses:
        '200':
          description: Message accepted for relay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostMessageResponse'
        '400':
          description: Invalid envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Recipient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Duplicate message id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /messages/poll:
    get:
      summary: Poll pending messages for a user
      operationId: pollMessages
      parameters:
        - in: query
          name: user_id
          required: true
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Pending messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollMessagesResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /messages/ack:
    post:
      summary: Acknowledge delivered messages
      operationId: ackMessages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AckMessagesRequest'
      responses:
        '200':
          description: Acknowledgement accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckMessagesResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

components:
  schemas:
    RegisterRequest:
      type: object
      additionalProperties: false
      required:
        - user_id
        - identity_key_ed25519
        - dh_key_x25519
        - nonce
        - signed_at
        - signature
      properties:
        user_id:
          type: string
          description: UUID/string identifier
        identity_key_ed25519:
          type: string
          description: Base64 Ed25519 public key
        dh_key_x25519:
          type: string
          description: Base64 X25519 public key
        nonce:
          type: string
          description: Base64 random 16-byte nonce
        signed_at:
          type: string
          format: date-time
        signature:
          type: string
          description: Base64 Ed25519 signature over canonical register string

    RegisterResponse:
      type: object
      additionalProperties: false
      required: [ok, user_id, registered_at]
      properties:
        ok:
          type: boolean
          const: true
        user_id:
          type: string
        registered_at:
          type: string
          format: date-time

    UserKeysResponse:
      type: object
      additionalProperties: false
      required:
        - user_id
        - identity_key_ed25519
        - dh_key_x25519
        - keyset_version
        - updated_at
      properties:
        user_id:
          type: string
        identity_key_ed25519:
          type: string
        dh_key_x25519:
          type: string
        keyset_version:
          type: integer
          minimum: 1
        updated_at:
          type: string
          format: date-time

    MessageEnvelopeRequest:
      type: object
      additionalProperties: false
      required:
        - version
        - message_id
        - from_user_id
        - to_user_id
        - sender_identity_key_ed25519
        - sender_dh_key_x25519
        - nonce
        - ciphertext
        - sent_at
        - signature
      properties:
        version:
          type: integer
          enum: [1]
        message_id:
          type: string
          format: uuid
        from_user_id:
          type: string
        to_user_id:
          type: string
        sender_identity_key_ed25519:
          type: string
        sender_dh_key_x25519:
          type: string
        nonce:
          type: string
        ciphertext:
          type: string
        sent_at:
          type: string
          format: date-time
        signature:
          type: string

    PostMessageResponse:
      type: object
      additionalProperties: false
      required: [ok, server_message_id, received_at]
      properties:
        ok:
          type: boolean
          const: true
        server_message_id:
          type: string
        received_at:
          type: string
          format: date-time

    PollMessageItem:
      type: object
      additionalProperties: false
      required: [server_message_id, from_user_id, received_at]
      properties:
        server_message_id:
          type: string
        from_user_id:
          type: string
        received_at:
          type: string
          format: date-time

    PollMessagesResponse:
      type: object
      additionalProperties: false
      required: [messages]
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/PollMessageItem'

    AckMessagesRequest:
      type: object
      additionalProperties: false
      required: [user_id, server_message_ids, acked_at]
      properties:
        user_id:
          type: string
        server_message_ids:
          type: array
          minItems: 1
          items:
            type: string
        acked_at:
          type: string
          format: date-time

    AckedMessageItem:
      type: object
      additionalProperties: false
      required: [server_message_id, envelope, received_at]
      properties:
        server_message_id:
          type: string
        envelope:
          $ref: '#/components/schemas/MessageEnvelopeRequest'
        received_at:
          type: string
          format: date-time

    AckMessagesResponse:
      type: object
      additionalProperties: false
      required: [ok, acked_count, acked_at, messages]
      properties:
        ok:
          type: boolean
          const: true
        acked_count:
          type: integer
          minimum: 0
        acked_at:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: '#/components/schemas/AckedMessageItem'

    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, request_id]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, retryable]
          properties:
            code:
              type: string
            message:
              type: string
            retryable:
              type: boolean
        request_id:
          type: string